Create Celebrities & Hobbies
In this query we are pulling all of the data from the who table and we are joining it up with everything in the hobbies table. We are doing this so the person who is sustaining this data will be able to understand who are all the celebrities in the data given, and also what their hobbies consist of.